backend/requirements.txt
fastapi
uvicorn
python-multipart
apscheduler
google-generativeai
numpy
psutil
